Updated valid json logic
In classes.py, self.last_trigger_time() used the last time
provided in the json for flare start, peak, end, and
last_data_time. For historical analysis, the start, peak, and
end may be filled out in the json because it is know, however
this was causing jsons to be erroneously identified as invalid
because the issue time would be earlier than the flare end time,
for example. Instead, the logic was changed to give precedence
to the last_data_time in the flare block as the appropriate
time to compare to the issue time.