Updated matching and metrics
In classes.py, changed the logic to identify the last eruption time and
last data time in last trigger time logic.
The last_data_time in the flare block is no longer being considered in
the last_eruption_time. Instead, the last_data_time now goes into the last_time.
The last_eruption_time is used to associated a forecast from a particular
flare and/or CME with an SEP event.
The last_time is compared to the issue time to ensure the forecast is valid.

In the metrics, removed the 100% multiplier from Symmetric Median Accuracy.
Decided to keep all percentages as ratios on the metrics files and will
multiple by 100% when making plots and reports.