Added more descriptions to metrics.py
Added more text to the commment blocks for each of the metrics and
skills scores in metrics.py. These include information about formula
and other notes about implementation that were not present in previous
versions.

Added a line to gitignore for the tests folder which currently houses
the ongoing unittest construction.

Other minor changes (add/delete empty lines of code)