Add TopDirectory feature to ID location of profile txts
SEP Scoreboard json files and intensity profile txt files are stored in
different directories, e.g. for iPATH and ASPECS.
The JSON files provide only the name of the time profile file and do
not indicate the directory in which they are found.
Rather than write exceptions into SPHINX for each SEP Scoreboard time profile
model, a new function was added to walk through the directory structure on the
users laptop and search for all txt files. When a txt file is found, the
name of the file is stored in a dictionary as the key and the associated
value is the file with its full path.

When running SPHINX, a new flag is now required: --TopDirectory "mydir/"

where mydir/ indicates the top level directory to go into to search for
time profile txt files. TopDirectory needs to be general enough to find
all txt files needed for the validation, but specific enough to avoid
searching through unnecessary directories or directories that may contain
copies of the same files.

--TopDirectory is a required input in this version.