Fix bug in Unmatching and time profile metrics
The unmatching algorithm didn't correctly set the end_time_match_status as
Unmatched, so end time, fluence, and time profile was still set as
SEP Event. This was corrected and all match statuses are now
set to Unmatched when a forecast is unmatched from an SEP event.

The above bug lead to the identification of a bug in the trimming
algorithm in the time_profile library. It was not checking for
times that were Pandas NaT and this lead to incorrect behavior
in time profile comparisons. This has been corrected.