Added new tests to metric metric_unittest
Added tests to get more coverage of metrics.py. Before this commit
coverage was 85 percent and is now at 99 percent. Only two lines in
metrics.py are not touched now, but I'm not entirely sure how to get to
it (part of a try/except but when I test it, the correct error is
raised).

Added new tests to  make sure garbage inputs are handled appropriately
and output the sys.exit or errors where appropriate.