Added last data time to issue time calculation
Added a class attribute called last_data_time_to_issue_time which takes
the very last time of the inputs/triggers ingested to make the forecast.
This is compared to the issue time and the difference is reported in minutes.
This time calculation tells how long it takes from the last data
ingested to the forecast being written out by the model. It includes
data availability, human-in-the-loop actions, and model process time.