Fixed match.py setup and matching indexing problems
After refactoring the observational criteria portion of match.py into two subroutines,
setup_match_all_forecasts and match_all_forecasts, some indexing problems became
apparent. setup_match_all_forecasts created the sphinx objects and saved them
to a dictionary organized by model and energy channel, matched_sphinx.
match_all_forecasts then went through the list of forecasts again and
aimed to pull out the appropriate sphinx object from matched_sphinx to add
more information. After the refactoring, I changed the forecast.valid()
routine to return a False value instead of a None value in the case that
a forecast had no trigger or input information at all.
A None allowed the forecast to be evaluated while a False specified that a
forecast should be skipped. the forecast.valid()
criteria was included in setup_match_all_forecasts, but not included in
match_all_forecasts, so this introduced an indexing problem that wasn't present
at the time of the original refactoring. To avoid mismatching logic in the two
subroutines, match_all_forecasts was modified to pull the sphinx objects directly
from the matched_sphinx dictionary and these objects were then updated with
additional matching criteria.

In debugging the above problem, a new .index attribute was added to the
forecast objects. This is a unique index for each forecast assigned
in the order that the forecast is read into SPHINX. This index can be used
to trace how a forecast is being handled within SPHINX.

Some comments and docstrings were updated for correctness.