Add all values relevant to matching to dataframe
Added boolean and timing information that is used in the matching
procedure. Previously, this information was reported to the log
using the match_report attribute in the SPHINX class. We
felt it was better to keep the matching information in the SPHINX
dataframe and avoid printing to screen or file.