Added check for duplicate entries in dataframe
Exact duplicate forecasts should not be allowed in processing
to avoid incorrectly inflated numbers. Forecasts are considered
duplicate if EVERY entry in the SPHINX_dataframe is exactly
the same. A hash is created for each row of the dataframe. If
multiple hashes are the same, then the duplicates are
removed and only one forecast remains. A warning with the
message DUPLICATE is printed out to the log.