Matching values to string; Explicit cast to timedelta; Change df syntax
All values related to matching that aren't used in validation are cast as
strings. They are saved in lists and this was causing problems with
read in/read out compatability. They are mainly for human reference,
so cast all to string to make SPHINX_dataframe maintenance easier.

In extract_time_forecast_type, the syntax:
sel_df.loc[len(sel_df)] = row.values

ended up changing the datatype of the columns in sel_df if
a None of NaT value was present in row. A warning indicated
that the previous behavior of this call was deprecated and
would act differently in the future. I think this introduced
a bug at this time. The next lines in the code expected a
column in sel_df to be in the format of datetime, but this
call cast it to a datatype of object instead. An attempt
to subtract two columns as a timedelta no longer cast to the
correct datatype, which caused a call of td.dt.total_seconds()
to crash. It was fixed by explicitly casting the td variable
to a timedelta.

Lines with a syntax setting a dataframe value as:
df['col1'].iloc[0] = value

were identified in a warning as deprecated in pandas 3.0, so
the syntax was changed as recommended in the warning.