Add --clean and improve logging for check_json.py (#114)
The --clean option writes a new --ModelList file with problematic
jsons commented out.

Now uses the logging module configured so that warnings from deeper
into the code go to the stdout with the rest of the output.

Added a new file state prefix, INVALID, for files that fail the
valid_forecast() test.  These files can still be processed by SPHINX,
and will not be commented out by --clean.