Add Threshold and SEP Start Time for match traceability
All thresholds applied to a given energy channel are saved to the
SPHINX_dataframe in a field called "All Thresholds in Prediction".
This reports all thresholds applied to a given energy channel within
the forecast. If a match status comes back as No Matching Threshold
(indicating that the threshold used to prepare the observations for
the given energy channel was not present in the predictions),
this field can show which thresholds were applied in the prediction.

Added a field called "All Threshold Crossing Times" to the dataframe.
This lists all observed threshold crossing times within the forecast
prediction window. This has no bearing on whether SPHINX decides a forecast
should be matched to those threshold crossings. It simply lists that
the thresholds were crossed at those times for matching traceability.