Fix for forecasts with no specified thresholds
If a forecast json contains a prediction for only peak_intensity,
peak_intensity_max, or sep_profile, there will be no thresholds
explicitly stated in the json file. This previously caused SPHINX
to skip the forecasts.

In the case of above, SPHINX now applies the thresholds used
to prepare the observations in order to complete its matching logic.
The forecasts are then evaluated following the normal workflow.

A warning is written to the log with the text "SUBSTITUTED THRESHOLDS".