Mismatch unit test
Got mismatch working. Small change in match.py in line 2252:
if cfg.do_mismatch and energy_key in cfg.mm_energy_key\
                and cfg.mm_model in fcast.short_name:

made it so that it checks if the energy key is in the mismatch
energy key, before it was energy_key == cfg.mm_energy_key.
Functionality is the same since equivalency would be accounted
for in the 'in' statement.