Ensure constant data types in dataframes
In the subroutines named like extract_flux_forecast_type, called
when First, Last, Mean, Max modes are used, there was a bug
related to the sel_df dataframes. The datatypes of df were not
being carried through to sel_df, which were returned from the subroutines.

This was addressed in two ways. First, sel_df was explicitly cast to have
the same datatypes as df.
Second, a changes in datatype was occurring when rows from df
were being added to sel_df using the syntax:
sel_df.loc[len(sel_df)] = row.values

row had a datatype of object, which changed the sel_df columns to datatype
of object. row was created using the syntax df.iloc[i], which resulted in a
Series and did not retain the datatype of each column. Instead, the syntax
row = df.iloc[[i]] was applied, which returns a single row dataframe.
This row was then added to sel_df using the pd.concat function. In this way,
all data types were preserved across functions, avoiding an unexpected change
of data type down the line.