Adding duplicate removal
The new library, duplicates.py, contains code to remove
duplicate forecasts entered into SPHINX via the ModelList.
Exact duplicates due to listing the same filename more than once
will be removed. Duplicate forecasts with exactly the same content,
event if they have different filenames and issue times, will be removed.
It is undesirable for models to issue duplicate forecasts with exactly
the same inputs, prediction windows, and predicted values.

A second duplicate removal occurs after the matching process is
complete and the SPHINX_dataframe is created. This was added a
while back. The subroutine to do this has been moved to duplicates.py.