Fix bug in removing duplicates
Duplicate forecasts were removed in reverse order of the
dup_indices list.  However, dup_indices was not guarunteed to be
sorted so a lower index could be removed before a higher index,
making an index out of range error possible.

Now the dup_indices are processed in reverse sorted order.