Improvements on handling thresholds and timing (#130)
 * Add Threshold and SEP Start Time for match traceability

All thresholds applied to a given energy channel are saved to the
SPHINX_dataframe in a field called "All Thresholds in Prediction".
This reports all thresholds applied to a given energy channel within
the forecast. If a match status comes back as No Matching Threshold
(indicating that the threshold used to prepare the observations for
the given energy channel was not present in the predictions),
this field can show which thresholds were applied in the prediction.

Added a field called "All Threshold Crossing Times" to the dataframe.
This lists all observed threshold crossing times within the forecast
prediction window. This has no bearing on whether SPHINX decides a forecast
should be matched to those threshold crossings. It simply lists that
the thresholds were crossed at those times for matching traceability.

 * Fix for forecasts with no specified thresholds (#4)

If a forecast json contains a prediction for only peak_intensity,
peak_intensity_max, or sep_profile, there will be no thresholds
explicitly stated in the json file. This previously caused SPHINX
to skip the forecasts.

In the case of above, SPHINX now applies the thresholds used
to prepare the observations in order to complete its matching logic.
The forecasts are then evaluated following the normal workflow.

A warning is written to the log with the text "SUBSTITUTED THRESHOLDS".

 * Add explicit check for thresholds for forecasts

The following predicted quantities should contain the thresholds
that were applied to determine those values:
            All clear
            Event lengths
            Fluence spectra
            Threshold crossings
            Probabilities

If thresholds are not present in those json blocks, SPHINX will
now explicitly skip those forecasts. A warning is printed to log
with the text "CORRUPT FORECAST" if a required threshold is missing.

 * Added unit tests for flux thresholds

Added two new unit tests to verify the following behaviors:

4_1) When a forecast JSON includes fields that require a flux threshold specification, but do not, the forecast is labeled "CORRUPT", and a flux threshold from the observation JSON is adopted by the forecast/SPHINX object.
4_2) When a forecast JSON does not include any fields that require a flux threshold specification, a flux threshold from the observation JSON is adopted by the forecast/SPHINX object.

Note that one deficiency of these tests is that they do not consider the case where there are multiple flux thresholds associated with a single energy threshold in the observation JSON.

 * Include flare last data time in eruption timing
 * Apply additional sort on Issue Time to dataframe

---------

Co-authored-by: lukestegeman <34717691+lukestegeman@users.noreply.github.com>