Improved duplicate removal (#135)
Fixes #126. A new library, duplicates.py, contains code to remove
duplicate forecasts entered into SPHINX via the ModelList.
Exact duplicates due to listing the same filename more than once
will be removed. Duplicate forecasts with exactly the same content,
even if they have different filenames and issue times, will be removed.
It is undesirable for models to issue duplicate forecasts with exactly
the same inputs, prediction windows, and predicted values.

The duplicate removal compares the resume SPHINX_dataframe to the forecasts read into SPHINX.
If the list of forecasts read in have an exact filename (Forecast Source) that is
already present in the resume SPHINX_dataframe, then remove that forecast from
the input list and do not process.  Another duplicate removal occurs after the matching process is
complete and the SPHINX_dataframe is created. This was added a
while back. The subroutine to do this has been moved to duplicates.py.

Forecasts that were duplicates of themselves in the ModelList are logged as DUPLICATE INPUT FORECAST.
Forecasts that were duplicates of the resume SPHINX_dataframe are logges as DUPLICATE RESUME FORECAST.
After the matching is completed and the SPHINX_dataframe is made from the SPHINX objects, any
duplicates are logged as DUPLICATE SPHINX FORECAST.

---------

Co-authored-by: Ricky Egeland <ricky.egeland@gmail.com>