Pass SPHINX dataframe directly to reporting module (#142)
Added return statements to sphinx.validate() and validation.intuitive_validation() to allow the passage of the SPHINX dataframe from the SPHINX validation module to the SPHINX reporting module. When running SPHINX's full validation protocol, the dataframe will be passed directly from the validation step to the reporting step. If the reporting module is run as a standalone code, the SPHINX dataframe will be read in from disk. Fixes #141.