Fix crash when removing resume duplicates
When both input duplicates and resume duplicates are present,
sphinx would crash because the first duplicate removal changed
the index order of the Forecast object array, which was saved
and reused in a dataframe for the resume duplicate removal.

Changed to regenerate the Forecast object dataframe in
remove_resume_duplicates() and remove existence of fcast_df
passed from remove_forecast_duplicates().