Shortname Grouping for UMASEP
UMASEP puts its submodules in its shortname, i.e. 'UMASEP-10 PCP',
which caused SPHINX to group these forecasts differently from
the rest of the model forecasts. This behavior is not how we want
to validate this model. So, the changes here are to make it such
that for UMASEP models, the submodule is no longer included in the
SPHINX validation. Most of the changes are to call a function written
in object_handler.py which rewrites the short_name string to be
just 'UMASEP' and whatever the energy channel is, i.e. 'UMASEP-10'.
There's also an additional line in the config file named
shortname_grouping that can be set to True/False to dictacte
if we group these UMASEP names together. If False, SPHINX should
behave as it did before.