Cleaning up
Cleaning up a bit and making sure that if the config line for
shortname_grouping is set to False then the previous behavior
is still working as it used to. Ran unittests to ensure those
results won't change based upon the contents of this PR, and they
passed (except the 1 match test that always fails).