Address issues #110 (logging) and #137 (resume)
For issue #110, added a try except block at the very top level in
bin/sphinx.py to capture exceptions in the log file rather that stdout.

For issue #137, moved the line to read in the resume dataframe before
reading in the forecast and observation jsons to avoid wasting
time when there is a problem with the resume dataframe.