Fixed bugs in previous commit as well as in old code in classes
Fixed some bugs in the previous large commit to get the code to run.

In classes.py, SPHINX attribute return_predicted_fluence_spectrum didn't
account for the possibility that the observations had not been
included in the SPHINX object. Now that there are sphinx objects created
from forecast objects that are not evaluated, there may not be any
observational information present. Added checking to account for that.
Also fixed in return_predicted_fluence.