New validation unittest (#157)
Improve robustness of validation unit tests             

Change assertEqual() tests for floats to assertAlmostEqual().  This fixes tests
that were not passing due to differences at the machine precision level.
                           
Added a new function in tests/utils.py that replaces the block of code that was
previously at the end of many of the steps - which matched the metric output
file to a test dictionary and compared the two to ensure the output files were
being made correctly. This block used to contain the try/except structure that
has now been generalized. This block also utilizes pandas.read_csv instead of
python's built in csv reader, which means the read in csv file now retains its
typing (floats are properly floats, strings are strings, etc.).  The test first
tries to see if both the output file entry and test dictionary are NaNs - if
they are both NaN it passes. Then if the output file entry is a string, it tests
the strings with an assertEquals, anything that hasn't been tested yet is
probably a float so it's tested with assertAlmostEquals.