Updated and generalized shortname changer
Made the shortname changer more generalized by using a list of tuples
in the config file that contain certain patterns and what to replace
the shortname with if the original shortname contains the pattern.
This function is called in three spots, match.py,
validation_json_handler.py  and classes.py. In  classes.py I replace
the shortname with the new one, but the original shortname
(i.e. UMASEP-10 PCP) is kept within the forecast object under the name
 of 'original_shortname' and is now part of the sphinx_dataframe
as a new column of Original Model Short Name (last column).
Any information that could be printed about the changing of the
shortname is accessible through debug level logging.