Updated Unit tests
Made a couple of changes to the unit tests to accomodate for the new
column in the dataframe (line added to dictionary in utils.py for
matching). Created new unit test that strictly tests the
shortname_grouper by changing the name of the model in the
json (Test_model_0) and patches the shortname_grouper in the mocked
config to change the name to 'new_shortname_for_testing'. The rest of
the test is fairly standard, and tests that this new shortname
is found in the dataframe.