Shortname shortener (#154)
* Generalized short name change capability added

* Shortname Grouping for UMASEP

UMASEP puts its submodules in its shortname, i.e. 'UMASEP-10 PCP',
which caused SPHINX to group these forecasts differently from
the rest of the model forecasts. This behavior is not how we want
to validate this model. So, the changes here are to make it such
that for UMASEP models, the submodule is no longer included in the
SPHINX validation. Most of the changes are to call a function written
in object_handler.py which rewrites the short_name string to be
just 'UMASEP' and whatever the energy channel is, i.e. 'UMASEP-10'.
There's also an additional line in the config file named
shortname_grouping that can be set to True/False to dictacte
if we group these UMASEP names together. If False, SPHINX should
behave as it did before.

Making sure that if the config line for
shortname_grouping is set to False then the previous behavior
is still working as it used to. Ran unittests to ensure those
results won't change based upon the contents of this PR, and they
passed (except the 1 match test that always fails).

* Updated and generalized shortname changer

Made the shortname changer more generalized by using a list of tuples
in the config file that contain certain patterns and what to replace
the shortname with if the original shortname contains the pattern.
This function is called in three spots, match.py,
validation_json_handler.py  and classes.py. In  classes.py I replace
the shortname with the new one, but the original shortname
(i.e. UMASEP-10 PCP) is kept within the forecast object under the name
 of 'original_shortname' and is now part of the sphinx_dataframe
as a new column of Original Model Short Name (last column).
Any information that could be printed about the changing of the
shortname is accessible through debug level logging.

* Updated Unit tests

Made a couple of changes to the unit tests to accomodate for the new
column in the dataframe (line added to dictionary in utils.py for
matching). Created new unit test that strictly tests the
shortname_grouper by changing the name of the model in the
json (Test_model_0) and patches the shortname_grouper in the mocked
config to change the name to 'new_shortname_for_testing'. The rest of
the test is fairly standard, and tests that this new shortname
is found in the dataframe.