Added bin/explore_json.py to create SEP lists
explore_json.py will read in a list of observation or forecast jsons
to extract SEP start and end times and print them to a list.

It is also intended to extract observation or prediction window
start and end times and check whether the time period covered
by these windows is continuous. This feature is not yet implemented.

This code is still under development.