Not eval unittest updates (#167)
* Updated Validation Unit Tests for SPHINX not eval

Updated unit tests to work with the new SPHINX workflow following
not_evaluated updates. Tests will pass, but new unit tests will
need to be built to actually test the not_evaluated output file, since
that is a new output file - but this is out of the scope of this update. 
(This part was already committed, but getting recommitted due to git bookkeeping issues.)

* Adding new column to dataframe in unittests

New column is SEP All Clear Probability Threshold