Forecast report ongoing event status not evaluated
The particle_intensities block in the SEP Scoreboard forecast jsons can contain
a field called ongoing_events where a model can report that it is aware that
an SEP event is in progress. Forecasts that indicate an event is ongoing
will be removed from the analysis so that they are neither credited or
penalized. These forecasts will be found in the SPHINX_removed dataframe.

The UMASEP model is currently the only model that does this.