Modify match algorithm for better outcomes
Flares with class >=M1 and longitude >= 20deg
are selected over flares with lower classes and more eastward longitudes.
If multiple flares >=M1 and >=20 deg are present, then the closest to the observed SEP
is chosen. If no flares satsify those conditions, then the flare
closest to the time of the SEP threshold crossing is chosen.

The preferred_cme subroutine takes a series of CME triggers as input.
CMEs with v >= 600km/s and half width >= 35 deg are selected. If multiple
CMEs satsify those conditions, then the closest to the observed SEP event is selected.
If the CMEs are narrow, but exceed 600 km/s, then the fastes CME is selected.
If all CMEs have v < 600km/s, then the CME closest in time to the
observed SEP event is selected.

Improved formatting in logging for clarity and compactness.