Improve unmatching logic (#174)
* Improve unmatch methodology

When multiple forecasts from the same model are matched to
the same observed SEP event, SPHINX attempts to identify the
correct forecast that should be associated with the event and
unmatch the other forecasts. This can only be done when the
forecast uses flare or CME trigger information.

The previous version of the unmatch feature simply compared
the flare or CME times and picked the forecast with the
eruption time closest to the time of observed SEP threshold crossing.
This logic was too simplistic and resulted in errors.

The matching logic has been updated to include a simple likelihood
analysis comparing CME speed and width and flare intensity and longitude
to pick out the CME or flare that should most likely be matched to
the SEP event. The logic is improved, but still simplistic and is meant
to be a placeholder until specific flare and CME associations can be specified
in the SPHINX observations.

The preferred_flare subroutine takes a series of
flares triggers as input.  Flares with class >=M1 and longitude >= 20deg
are selected over flares with lower classes and more eastward longitudes.
If multiple flares >=M1 and >=20 deg are present, then the closest to the observed SEP
is chosen. If no flares satsify those conditions, then the flare
closest to the time of the SEP threshold crossing is chosen.

The preferred_cme subroutine takes a series of CME triggers as input.
CMEs with v >= 600km/s and half width >= 35 deg are selected. If multiple
CMEs satsify those conditions, then the closest to the observed SEP event is selected.
If the CMEs are narrow, but exceed 600 km/s, then the fastest CME is selected.
If ALL CMEs are fast and wide, then preferred_cme returns and empty array
so that selection can default to a flare trigger, if present. If no flare
triggers are present, then the CME closest in time to the
observed SEP event is selected.

Improved robustness of flare timing checking. Flare start time and flare last data time
are now checked because the start time field is optional and may not be present.

Changed logic in classes.py last_trigger_time() to use flare start time, if present,
then to defer to peak_time (was previously switched).