Finalize all clear grid; correct flare analysis
In classes.py Forecast class, a line of code checked if
flare last_data_time was equal to None rather than checking
pd.isnull(). This line of code was not updated when we updated
null file formats. This means that forecasts providing only
flare last data time and not start, peak, end were not being
identified as eruption-triggered models and didn't have unmatching
applied. SAWS-ASPECS was most affected and is now treated appropriately.

In match.py, changed timing comparisons to be inclusive of end time
(previously was exclusive of end time). A SEPSTER2D SEPVAL challenge
forecast happened to have a prediction window that ended on the very
same minute that an SEP threshold was crossed, so it was getting
identified as a miss instead of a hit. Changed logic so that the
last minute of the prediction window is included in the evaluation.

Modified the unmatch logic slightly. Changed to include strong flares
and/or CMEs within 1.5 hours of the best flare/CME. This choice was
driven by March 7, 2012, which had two X-class flares within an
hour and 15 minutes of each other. So close together, that difficult
to determine which is the one to be associated to the SEP event
without a historical analysis. Forecasts from strong triggers so close
together are chosen to all be associated to a following SEP event.

In preferred_flare() in match.py, changed logic to:
- Check for western X-class flares, choose one closest to the SEP event
- If no webstern flares, choose closest X-class flare to SEP event
- If no X-class flares, choose closest M-class flare to SEP event
- If not M-class flares, take closest flare to SEP event

Create All Clear Grid:
Wrote all_clear_grid subroutine in post_analysis. It accepts a list of models
and SEP event and non-event date periods, then produces a single answer
for each period:
any Hit = Hit
all Misses = Miss
any False Alarm = False Alarm
all Correct Negatives = Correct Negative
This is similar to the previously implemented deoverlapping code and the two
will be combined for constency in a future update.
The subroutine recalculates All Clear metrics and also produces a csv file
grid with SEP event or non-event period on the vertical axis and models on the
horizontal axis. The cells are filled in with Hit/Miss for SEP events or FA/CN
for non-events. This can be read into excel to format and create a visual
representation of SEP performance across models for each event.

Note that the config has been changed to apply mismatch to UNSPELL.