Deoverlap Probability in post_analysis.py
Added ability to deoverlap probability forecasts by selecting
the maximum value of the forecasted probabilities in a given time window.
The maximum probability is found for all forecasts associated
with an SEP event (matches was is already produced in the
probability_selections_*_Max files) and identifies the maximum
probability for non-event periods. The non-event periods may
be specified in an input file (e.g. SEPVAL challenge) or created
as regular time intervals between the first and last model forecasts,
e.g. daily. The deoverlapping works the same as All Clear, but
selects max probability rather than identifying a hit, miss, etc.