AWT Percent Error now works
Got this bit of code working as intended. AWTPE can be interpreted as
a weighted AWT accounting for the amount of time between the model's
inputs and the threshold crossing.
AWTPE = AWT / (Time between last input and obs threshold crossing)

Looking at this maybe percent error is the wrong nomenclature to use
here. Maybe just weighted AWT could be appropriate, but I will let
Ricky and Katie decide that