Add metrics to flux-related selections files; fix mismatch bug (#183)
* Add metrics to flux-related selections files; fix mismatch bug

In the mismatch workflow, REleASE forecasts were getting thrown out
to the SPHINX_removed file. This was because the REleASE json fields are
all strings but the energy channels specified in config.py have float fields.
When the energy channels were being compared to each other, they were not
exactly the same and the forecasts were getting discarded due to mm_pred_energy_channel
not matching any of the energy channels read in from the json.
Added conversion of energy channel and threshold fields to float very early in
the process of reading in the json files to fix this issue.

Added the ratio as a metric for flux and fluence-related forecasts.

Added columns in the _selections_ files with metrics for each individual
observation-forecast pair. This will allow users to readily look at the
output files and perform filtering or identify outliers, etc.

In the process of adding the metrics to the _selections_ files, added logic that
removes all observed and predicted values <= 0 as we don't want them included in
the metrics. Any case where this occurs will no longer be written out to the
_selected_ files as they were in the past. Now, only the observation-forecast
pairs that went into the metrics calculations will be present in the _selections_ files.

Commented out/removed some fields from the AWT metrics files because they
were too redundant. All fields that related a time to "Observed SEP Start Time"
were commented because, in the current way we use SPHINX, this is always the
same as "Observed SEP Threshold Crossing Time". If there is ever a case when
these two times may differ, then it would be worthwhile to add the fields
back in.

* Fixed test_validation.py (#195)

---------

Co-authored-by: Clayton Allison <68346264+claya71@users.noreply.github.com>