Update time profile metrics
When comparing the observed and predicted time profiles, compare time profiles only when both are above threshold.

Metrics for the time profiles are calculated by taking a point-by-point comparison
and calculating the mean of the metric across the time profile. For multiple
time profiles, get a final metric, but taking the mean or median of the mean
metric across the profile.

Changed so that, if the endpoint metric was a mean, the mean metric was calculated
across each time profile and then the mean was taken of the means.
If the endpoint metric was a median, then the median metric was calcualted across each
time profile and the median was taken of the medians.

So the metrics are mean all the way through or median all the way through. This included adding
new columns and renaming previous columns in the time_profile_selections_* files with:
- Ratio --> Mean Ratio
- Median Ratio
- Error --> Mean Error
- Median Error
- Absolute Error --> Mean Absolute Error
- Median Absolute Error
- Log Error --> Mean Log Error
- Median Log Error
- Absolute Log Error --> Mean Absolute Log Error
- Median Absolute Log Error